Migrate mocks to TS
