+++ tests/mocks/mockInstabug.ts
