+++ tests/mocks/mockParseErrorStackLib.ts
