+++ tests/mocks/mockXhrNetworkInterceptor.ts
